pyproject.toml
[project]
name = "gpx2mesh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipympl>=0.9.7",
    "matplotlib>=3.10.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scipy>=1.15.3",
    "trimesh[easy]>=4.6.11",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.1",
    "ruff>=0.12.5",
]
